Issue #45: Path relative to config.yml
The config.yml of each repo contains information about the location of
the directories associated to it (output, services and templates). It is
assumed that these directories are relative to the cwd of the process
running saasherder, however this does not make sense when using
saasherder python libraries.

Since the path to the config file can be specified as an argument, this
changes makes the directories relative to the path of the config file,
not to the cwd of the process using the saasheder library.