Issue #45: validation scopes (#59)
* Issue #45: Reindent cli.py and saasherder.py

Use 4 spaces instead of 2

* Issue #45: Add ContainerRequestsLimitsRule

This commit adds the necessary changes to cli.py and saasherder.py to
perform generic validation checks.

A new command `saasherder validate  [--context CONTEXT]` is now
available.

One validation rule, ContainerRequestsLimitsRule, has been implemented.

* Issue #45: Path relative to config.yml

The config.yml of each repo contains information about the location of
the directories associated to it (output, services and templates). It is
assumed that these directories are relative to the cwd of the process
running saasherder, however this does not make sense when using
saasherder python libraries.

Since the path to the config file can be specified as an argument, this
changes makes the directories relative to the path of the config file,
not to the cwd of the process using the saasheder library.

* Issue #45: Use git -C to run git commands

Instead of running cd <dir> && git ...

* Issue #45: Refactor changelog/saasherder deps

New Changelog objects need an associated Saasherder as an
argument

* Issue #45: validation returns a list of errors

* Issue #45: Report validation only if hash defined

Skip the services that have `hash: None` in the service yaml, as
those will be skipped by SaasHerder when deploying to production.